Moved previously added functions into utils
The code parsing and env var collecting funtions added in the previous commit have been
moved into a seperate until.ts file for static access. The functions have also been
seperated into distinct functions and are no longer codependent.
